Moves moveViews command - re: #110583
